bookshelf/requirements.txt
flask
flask-sqlalchemy
python-dotenv
pytest
transformers
torch
PyJWT
